+++ localize.h
@@ -8,7 +8,7 @@ static char LOCALIZE_INIT;
 
 void localize_init();
 void localize_update();
-void localize_calculate(int* data);
+void localize_calculate(uint16_t* data);
 float* localize_location();
